Move app to its own file
